Some fixes and changing the way that picture type is inferred from a structure
